move nonkube components to the correct directory
